refactor: remove useSiweNextAuth Hook, move code into provider
