--- packages/rainbowkit-siwe-next-auth/src/useSiweNextAuth.ts
@@ -1,85 +1,0 @@
-import {
-  AuthenticationConfig,
-  createAuthenticationAdapter,
-} from '@rainbow-me/rainbowkit';
-import { getCsrfToken, signIn, signOut, useSession } from 'next-auth/react';
-import { useMemo } from 'react';
-import { SiweMessage } from 'siwe';
-
-type UnconfigurableMessageOptions = {
-  address: string;
-  chainId: number;
-  nonce: string;
-};
-
-type ConfigurableMessageOptions = Partial<
-  Omit<SiweMessage, keyof UnconfigurableMessageOptions>
-> & {
-  [Key in keyof UnconfigurableMessageOptions]?: never;
-};
-
-export type GetSiweMessageOptions = () => ConfigurableMessageOptions;
-
-export interface UseSiweNextAuthOptions {
-  getSiweMessageOptions?: GetSiweMessageOptions;
-}
-
-export function useSiweNextAuth({
-  getSiweMessageOptions,
-}: UseSiweNextAuthOptions = {}): AuthenticationConfig<SiweMessage> {
-  const { status } = useSession();
-  const adapter = useMemo(
-    () =>
-      createAuthenticationAdapter({
-        createMessage: ({ address, chainId, nonce }) => {
-          const defaultConfigurableOptions: ConfigurableMessageOptions = {
-            domain: window.location.host,
-            statement: 'Sign in with Ethereum to the app.',
-            uri: window.location.origin,
-            version: '1',
-          };
-
-          const unconfigurableOptions: UnconfigurableMessageOptions = {
-            address,
-            chainId,
-            nonce,
-          };
-
-          return new SiweMessage({
-            ...defaultConfigurableOptions,
-
-            // Spread custom SIWE message options provided by the consumer
-            ...getSiweMessageOptions?.(),
-
-            // Spread unconfigurable options last so they can't be overridden
-            ...unconfigurableOptions,
-          });
-        },
-
-        getMessageBody: ({ message }) => message.prepareMessage(),
-
-        getNonce: async () => {
-          const nonce = await getCsrfToken();
-          if (!nonce) throw new Error();
-          return nonce;
-        },
-
-        signOut: async () => {
-          await signOut({ redirect: false });
-        },
-
-        verify: async ({ message, signature }) => {
-          const response = await signIn('credentials', {
-            message: JSON.stringify(message),
-            redirect: false,
-            signature,
-          });
-
-          return response?.ok ?? false;
-        },
-      }),
-    [getSiweMessageOptions]
-  );
-
-  return { adapter, status };
-}
